+++ view/frontend/web/js/view/payment/method-renderer/hgw-directdebit.js
@@ -42,8 +42,6 @@ define(
                             }
                         )
                     }
-                } else {
-                    console.log('Customer is not logged in.');
                 }
 
                 return this;